refactor: Move create_db_and_tables module from vectors to pgvector
Moved create_db_and_tables module to pgvector so it only handles pgvector related db creation

Refactor #COG-170
