+++ cognee/api/v1/add/add_v2.py
@@ -4,11 +4,11 @@ from cognee.modules.users.methods import get_default_user
 from cognee.modules.pipelines import run_tasks, Task
 from cognee.tasks.ingestion import save_data_to_storage, ingest_data
 from cognee.infrastructure.databases.relational import create_db_and_tables as create_relational_db_and_tables
-from cognee.infrastructure.databases.vector import create_db_and_tables as create_vector_db_and_tables
+from cognee.infrastructure.databases.vector.pgvector import create_db_and_tables as create_pgvector_db_and_tables
 
 async def add(data: Union[BinaryIO, list[BinaryIO], str, list[str]], dataset_name: str = "main_dataset", user: User = None):
     await create_relational_db_and_tables()
-    await create_vector_db_and_tables()
+    await create_pgvector_db_and_tables()
 
     if user is None:
         user = await get_default_user()
