+++ cognee/infrastructure/databases/vector/__init__.py
@@ -4,4 +4,3 @@ from .models.CollectionConfig import CollectionConfig
 from .vector_db_interface import VectorDBInterface
 from .config import get_vectordb_config
 from .get_vector_engine import get_vector_engine
-from .create_db_and_tables import create_db_and_tables
