+++ cognee/infrastructure/databases/vector/pgvector/__init__.py
@@ -1,1 +1,2 @@
 from .PGVectorAdapter import PGVectorAdapter
+from .create_db_and_tables import create_db_and_tables
